+++ R/utils.r
@@ -4,6 +4,14 @@ assert_gh <- function() {
     invisible(NULL)
 }
 
+assert_string <- function(x) {
+    if (!is.character(x) || length(x) != 1L) {
+        nm <- deparse(substitute(x))
+        cli::cli_abort("`{nm}` must be a string.")
+    }
+    invisible(NULL)
+}
+
 has_gh <- function() nzchar(gh_cmd())
 
 gh_cmd <- function() Sys.which("gh")