+++ NEWS.md
@@ -5,3 +5,4 @@ New features
 ------------
 
 * `gh_browse()` opens up a GitHub repository in a web browser (#4)
+* `gh_repo_edit()` edits repository settings (#6)